requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
pillow==10.2.0
python-dotenv==1.0.0